+++ createllm/tests/test_createllm.py
@@ -1,0 +1,21 @@
+#!/usr/bin/env python
+
+"""Tests for `createllm` package."""
+
+
+import unittest
+
+from createllm import createllm
+
+
+class TestCreatellm(unittest.TestCase):
+    """Tests for `createllm` package."""
+
+    def setUp(self):
+        """Set up test fixtures, if any."""
+
+    def tearDown(self):
+        """Tear down test fixtures, if any."""
+
+    def test_000_something(self):
+        """Test something."""